Consolidated RevisionDataOutput implementations into a single class and made them more efficient.
